PIM-4778: move basic implementation of searchable repository
