+++ src/Pim/Bundle/ClassificationBundle/Doctrine/ORM/Repository/TagRepository.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\ClassificationBundle\Doctrine\ORM\Repository;
 
-use Doctrine\ORM\EntityRepository;
+use Akeneo\Bundle\StorageUtilsBundle\Doctrine\ORM\Repository\SearchableRepository;
 use Pim\Component\Classification\Repository\TagRepositoryInterface;
 
 /**
@@ -12,7 +12,7 @@ use Pim\Component\Classification\Repository\TagRepositoryInterface;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class TagRepository extends EntityRepository implements TagRepositoryInterface
+class TagRepository extends SearchableRepository implements TagRepositoryInterface
 {
     /**
      * {@inheritdoc}
@@ -48,42 +48,6 @@ class TagRepository extends EntityRepository implements TagRepositoryInterface
         }
 
         return $codes;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function findBySearch($search = null, array $options = [])
-    {
-        if (method_exists($this, 'getAlias')) {
-            $alias = $this->getAlias();
-        } else {
-            $alias = 'alias';
-        }
-
-        $qb = $this->createQueryBuilder($alias);
-        $qb->select("$alias.id, $alias.code");
-
-        if (null !== $search) {
-            $qb->where("$alias.code like :search")->setParameter('search', "%$search%");
-        }
-
-        if (isset($options['limit'])) {
-            $qb->setMaxResults((int) $options['limit']);
-            if (isset($options['page'])) {
-                $qb->setFirstResult((int) $options['limit'] * ((int) $options['page'] - 1));
-            }
-        }
-
-        $results = [];
-        foreach ($qb->getQuery()->getArrayResult() as $row) {
-            $results[] = [
-                'id'   => $row['code'],
-                'text' => $row['code']
-            ];
-        }
-
-        return $results;
     }
 
     /**
